% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papisr.r
\docType{package}
\name{papisr-package}
\alias{papisr}
\alias{papisr-package}
\title{papisr: Bundle of convenience functions for papis workflows in R.}
\description{
Provides some convenience functions for [https://github.com/papis/papis][papis] workflows in R. Papis is '[p]owerful and highly extensible command-line based document and bibliography manager'. The package does not actually require `papis` to be installed in order for its functions to work.
}
\details{
Bundle of convenience functions for papis workflows in R.
Provides some convenience functions for \link[=papis]{https://github.com/papis/papis} workflows in R. Papis is '\link{p}owerful and highly extensible command-line based document and bibliography manager'. The package does not actually require \code{papis} to be installed in order for its functions to work.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stasvlasov.github.io/papisr/}
  \item Report bugs at \url{https://github.com/stasvlasov/papisr/issues}
}

}
\author{
\strong{Maintainer}: Stanislav Vlasov \email{s.vlasov@uvt.nl}

}
\keyword{internal}
