% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papisr.r
\docType{package}
\name{papisr-package}
\alias{papisr}
\alias{papisr-package}
\title{papisr: papisr - R workflow for tabulating papis records}
\description{
Provides some convenience functions for [papis](https://github.com/papis/papis) workflows in R. Papis is a 'powerful and highly extensible command-line based document and bibliography manager'. The package does not actually require `papis` to be installed in order for its functions to work.
}
\details{
Bundle of convenience functions for papis workflows in R.

Provides some convenience functions for \href{https://github.com/papis/papis}{papis} workflows in R. Papis is a 'powerful and highly extensible command-line based document and bibliography manager'. The package does not actually require \code{papis} to be installed in order for its functions to work.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stasvlasov/papisr}
  \item Report bugs at \url{https://github.com/stasvlasov/papisr/issues}
}

}
\author{
\strong{Maintainer}: Stas Vlasov \email{s.vlasov@uvt.nl}

}
