% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papisr.r
\name{tabulate_papis_records}
\alias{tabulate_papis_records}
\title{Tablulate papis records}
\usage{
tabulate_papis_records(papis_records, ..., use_path_as_row_names = FALSE)
}
\arguments{
\item{papis_records}{List of papis records as returned by \code{collect_papis_records()}}

\item{...}{Colums specification as named expressions that are evaluated in papis record environment where two variables are bound - \code{path} and \code{info} (see \code{collect_papis_records()} for details)}

\item{use_path_as_row_names}{Use 'path' or papis record as a row's name}
}
\value{
Matrix
}
\description{
Tablulate papis records
}
